CLI: refactored CLI options - Moved click options and parsers to cli_utils.py
