+++ soprano/scripts/cli.py
@@ -24,9 +24,7 @@ __date__ = "July 04, 2022"
 
 
 import click
-import os
 from soprano.scripts import  nmr, nmr_plot
-from configparser import ConfigParser
 import logging
 import click_log
 # logging
@@ -41,46 +39,16 @@ help_text = """
 A CLI tool to streamline common soprano tasks. It has various 
 subcommands, each of which has its own set of options and help.
 """
-# join home and config file
-home = os.path.expanduser('~')
-# get default soprano config file:
-DEFAULT_CFG = os.environ.get('SOPRANO_CONFIG', f'{home}/.soprano/config.ini')
-
-#callback to load config file
-def configure(ctx, param, filename):
-    cfg = ConfigParser()
-    cfg.read(filename)
-    ctx.default_map = {}
-    for sect in cfg.sections():
-        command_path = sect.split('.')
-        if command_path[0] != 'soprano':
-            continue
-        defaults = ctx.default_map
-        for cmdname in command_path[1:]:
-            defaults = defaults.setdefault(cmdname, {})
-        defaults.update(cfg[sect])
 
 @click.group(
     name="Soprano Command Line Interface",
     help=help_text, epilog=epilog,
     invoke_without_command=True)
     
-@click.option(
-    '-c', '--config',
-    type         = click.Path(dir_okay=False),
-    default      = DEFAULT_CFG,
-    callback     = configure,
-    is_eager     = True,
-    expose_value = False,
-    show_default = True,
-    help         = 'Read option defaults from the specified INI file'
-                    'If not set, first checks environment variable: '
-                    '``SOPRANO_CONFIG`` and then ``~/.soprano/config.ini``',
-)
+
 @click_log.simple_verbosity_option(logger)
 
 def soprano():
-    logger.debug('Testing my logger')
     pass
 
 soprano.add_command(nmr.nmr)
