+++ soprano/scripts/nmr.py
@@ -30,23 +30,27 @@ __date__ = "July 08, 2022"
 
 import click
 import numpy as np
-import re
-import os
-import sys
-import re
 from ase import io
 from ase.visualize import view as aseview
 from ase.units import Ha, Bohr
 from soprano.properties.labeling import UniqueSites, MagresViewLabels
 from soprano.properties.nmr import *
-from soprano.data.nmr import _el_iso, _get_isotope_list
+from soprano.data.nmr import _get_isotope_list
 from soprano.selection import AtomSelection
-from soprano.utils import has_cif_labels, average_quaternions
+from soprano.utils import has_cif_labels
 import pandas as pd
 import warnings
-from collections import OrderedDict, defaultdict
 import logging
 import click_log
+from soprano.scripts.cli_utils import \
+                                    add_options,\
+                                    NMREXTRACT_OPTIONS, \
+                                    average_quaternions_by_tags,\
+                                    get_missing_cols,\
+                                    get_matching_cols,\
+                                    print_results,\
+                                    find_XHn_groups,\
+                                    sortdf
 # logging
 logging.captureWarnings(True)
 logger = logging.getLogger('cli')
@@ -70,126 +74,6 @@ cif2cell. e.g. for CASTEP:
 cif2cell mystructure.cif --export-cif-labels -p castep
 
 '''
-# We need to rename the columns to include units before printing
-UNITS = {
-    "MS_shielding": "ppm",
-    "MS_shift": "ppm",
-    "MS_anisotropy": "ppm",
-    "MS_reduced_anisotropy": "ppm",
-    "MS_span": "ppm",
-    "MS_alpha": "deg",
-    "MS_beta": "deg",
-    "MS_gamma": "deg",
-    "EFG_Vzz": 'Vm^-2',
-    "EFG_quadrupolar_constant": 'MHz',
-    "EFG_alpha": "deg",
-    "EFG_beta": "deg",
-    "EFG_gamma": "deg",
-    "EFG_NQR": "MHz",
-}
-
-implemented_properties = ['ms', 'efg']
-def isotope_selection(ctx, parameter, isotope_string):
-    """Parse the isotope string.
-    Args:
-        ctx: click context
-        parameter: click parameter
-        isotope_string (str): The isotopes specification, in the form ``"2H,15N" for deuterium and 15N``.
-    Returns:
-        dict: The isotope for each element specified. Formatted as::
-            {Element: Isotope}.
-    
-    """
-    if isotope_string == '':
-        return {}
-    isotope_dict = {}
-    for sym in isotope_string.split(","):
-        try:
-            el, isotope = _el_iso(sym)
-        except Exception as e:
-            raise e
-        isotope_dict[el] = isotope
-    return isotope_dict
-def keyvalue_parser(ctx, parameter, value):
-    """Parse strings in the form 'C:1,H:2' into a dictionary.
-        Also accept = as the separator between key and value.
-        e.g. the MS shift reference and gradient strings.
-        If the value is a single float, that will returned instead of a dict.
-    Args:
-        ctx: click context
-        parameter: click parameter
-        value (str): The references specification, in the form ``"C:100,H:123"``.
-                     If value is a single float, that will returned instead of a dict.
-    Returns:
-        dict: The values for each key specified. Formatted as::
-            {key: value}.
-    """
-    try:
-        return float(value)
-    except ValueError:
-        pass
-
-    keyvaluedict = {}
-    if value != '':
-        for sym in re.split(',', value):
-            try:
-                el, reference = re.split(":|=", sym)
-                keyvaluedict[el] = float(reference)
-            except Exception as e:
-                raise e
-    return keyvaluedict
-
-def get_column_list(ctx, parameter, value):
-    """Parse the column names string.
-    TODO: Document this a bit better.
-    Args:
-        ctx: click context
-        parameter: click parameter
-        value (str): The column names, comma-separated.
-                    Some shortcuts defined for MS_angles and EFG_angles.
-    Returns:
-        list: The column names specified.
-    """
-    if value == '' or value is None:
-        return None
-    # shortcuts for some column groups
-    special_names = {
-                    'minimal': 
-                        ['MS_shielding',
-                        'MS_anisotropy',
-                        'EFG_quadrupolar_constant',
-                        'EFG_asymmetry'
-                        ],
-                    'MS_defaults':
-                        ['MS_shielding',
-                        'MS_anisotropy',
-                        'MS_reduced_anisotropy',
-                        'MS_asymmetry'],
-                    'EFG_defaults':
-                        ['EFG_Vzz',
-                        'EFG_quadrupolar_constant',
-                        'EFG_asymmetry'],
-                    'MS_angles': 
-                        ['MS_alpha',
-                        'MS_beta',
-                        'MS_gamma'],
-                    'EFG_angles':
-                        ['EFG_alpha',
-                        'EFG_beta',
-                        'EFG_gamma',
-                            ],
-                    }
-    special_names['default'] = special_names['MS_defaults'] + special_names['EFG_defaults']
-
-    specified_columns = [c.strip() for c in value.split(',')]
-    # replace special names
-    for special_name, special_cols in special_names.items():
-        if special_name in specified_columns:
-            specified_columns.remove(special_name)
-            specified_columns.extend(special_cols)
-    # make sure no duplicates, preserving order
-    specified_columns = list(OrderedDict.fromkeys(specified_columns))
-    return specified_columns
 
 @click_log.simple_verbosity_option(logger)
 
@@ -200,173 +84,8 @@ def get_column_list(ctx, parameter, value):
                 nargs=-1,
                 type=click.Path(exists=True),
                 required=True)
-# option to select a subset of atoms
-@click.option('--select',
-                '-s',
-                'selection',
-                type=str,
-                default=None,
-                help='Selection string of sites include. e.g. \n'
-                '``-s C`` for only and all carbon atoms,\n'
-                '``-s C.1-3,H.1.2`` for carbons 1,2,3 and hydrogens 1 and 2,\n'
-                '``-s C1,H1a,H1b`` for any sites with the labels C1, H1a and H1b.'
-                )
-@click.option('--output',
-            '-o',
-            type=click.Path(exists=False),
-            default=None,
-            help='Output file name. If not specified, output is printed to stdout.')
-@click.option('--output-format',
-            '-f',
-            default=None,
-            type=click.Choice(['csv', 'json']),
-            help='Output file format. '
-            'If not specified, the format is guessed from output filename extension.')
-# merge output files
-@click.option('--merge',
-            '-m',
-            is_flag=True,
-            default=False,
-            help="If present, merge all files into a single output file.")
 
-# Option to specify the isotopes to use
-@click.option('--isotopes',
-            '-i',
-            callback = isotope_selection,
-            default='',
-            metavar = 'ISOTOPES',
-            help='Isotopes specification (e.g. ``-i 13C`` for carbon 13 '
-        '``-i 2H,15N`` for deuterium and 15N). '
-        'When nothing is specified it defaults to the most common NMR active isotope.')         
-# flag option to reduce by symmetry
-@click.option('--reduce',
-            '-r',
-            is_flag=True,
-            default=False,
-            help="Reduce the output by symmetry-equivalent sites. "
-        "The merged equivalent sites are combined according to ``--combine-rule``. "
-        "To see the rules used for each column, set ``--verbose``. "
-        "If there are CIF-style labels present, then these override the symmetry-grouping in "
-        "case of a clash. "
-        "Note that this doesn't take into account magnetic symmetry!")
-# symprec flag
-@click.option('--symprec',
-            type=click.FLOAT,
-            default=1e-4,
-            help="Symmetry precision for symmetry reduction. "
-        "Defaults to 1e-4.")
-# option to specify group_pattern for averaging
-@click.option('--average-group',
-            '-g',
-            type=str,
-            default=None,
-            help="Group pattern for averaging. "
-            "Currently only works for XHn groups such as CH3, CH2, NH2 etc. "
-            "You can specify several, comma separated as in ``-g CH3,CH2,NH2``. "
-            "If not specified, no averaging is performed.")
-# combine rule
-@click.option('--combine-rule',
-            default='mean',
-            type=click.Choice(['mean', 'first']),
-            help="How to combine the data from equivalent sites. "
-        "``mean`` is the default, which averages the data. "
-        "``first`` Takes the first item from each group of equivalent sites. "
-        "Special handling for labels, indices, tags and multiplicity. "
-        "Set verbose to True to see what rules have been used.")
-#  what to extract/analyse
-@click.option('--properties',
-            '-p',
-            type=click.Choice(implemented_properties),
-            default=['efg', 'ms'],
-            multiple=True,
-            help="Properties for which to extract and summarise e.g. ``-p ms.`` "
-            "They can be combined by using the flag multiple times: ``-p ms -p efg.`` "
-            "Defaults to both ms and efg.")
-# optional argument for the precision of the output
-@click.option('--precision',
-            type=click.INT,
-            default=3,
-            help="Precision of the output (number of decimal places). Defaults to 3.")
-# choose between Euler angle conventions 'zyz' or 'zxz'
-@click.option('--euler',
-            'euler_convention',
-            type=click.Choice(['zyz', 'zxz']),
-            default='zyz',
-            help="Convention for Euler angles. Defaults to ``zyz``.")
-# sort by df column
-@click.option('--sortby',
-            type=str,
-            default=None,
-            help="Sort by column. Defaults to sorting by site number. "
-            "It can be any column in the output. "
-            "For example ``--sortby EFG_Vzz``")
-@click.option('--sort-order',
-            type=click.Choice(['ascending', 'descending']),
-            default='ascending',
-            help="Sort order. Defaults to ascending.")
-# dictionary option to specify reference for each element
-@click.option('--references',
-            callback = keyvalue_parser,
-            default='',
-            help="Reference shielding for each element (in ppm). "
-            "The format is ``--references C:170,H:123``. "
-            "If the value is a single float, that reference will be used for all sites (not recommended!). ")
-@click.option('--gradients',
-            callback = keyvalue_parser,
-            default='',
-            help="Reference shielding gradients for each element. "
-            "Defaults to -1 for all elements. Set it like this: "
-            "``--gradients H:-1,C:-0.97``. "
-            "If the value is a single float, that gradient will be used for all sites (not recommended!). "
-            )
-# todo: have an option to set a file/env variable for the references... 
-# flag to include certain columns only
-@click.option('--include',
-            callback=get_column_list,
-            default=None,
-            help="Include only certain columns. "
-            "The columns are specified as a comma-separated list. "
-            "For example ``--include MS_shielding,EFG_Vzz``. "
-            "Defaults to all columns.")
-# flag to exclude certain columns
-@click.option('--exclude',
-            callback=get_column_list,
-            default=None,
-            help="Exclude certain columns. "
-            "The columns are specified as a comma-separated list. "
-            "For example ``--exclude MS_alpha,MS_beta,MS_gamma``. "
-            "Defaults to None.")
-# flag to filter results
-@click.option('--query',
-            type=str,
-            default=None,
-            help="Filter results based on query. "
-            "The filter is specified as a pandas query. "
-            "Note that you must enclose the query in quotes! "
-            "Refer to the column names without the units.  "
-            "For example ``--query 'MS_shielding > 100'``. "
-            "You can combine queries with ``and`` and ``or`` etc. "
-            "e.g. ``--query 'MS_shielding > 100 and MS_shielding < 180'``. "
-            "Defaults to #nofilter :).")
-
-# flag to view
-@click.option('--view',
-            is_flag=True,
-            default=False,
-            help="If present, view the structure(s) with the ASE GUI."
-            "Note that the ASE GUI can color the sites according to their tags. "
-            "This can be used to see what sites were tagged as equivalent.")
-
-
-# quiet flag
-@click.option('--quiet',
-            '-q',
-            is_flag=True,
-            default=False,
-            help="If present, log less information.")
-
-      
-
+@add_options(NMREXTRACT_OPTIONS)
 
 def nmr(files,
         selection,
@@ -599,17 +318,6 @@ def nmr(files,
 
         # what columns should we include/exclude?
         essential_columns = ['labels', 'species', 'multiplicity', 'tags', 'file']
-        # shortcuts for some column groups
-        special_names = {'MS_angles': 
-                            ['MS_alpha',
-                            'MS_beta',
-                            'MS_gamma'],
-                        'EFG_angles':
-                            ['EFG_alpha',
-                            'EFG_beta',
-                            'EFG_gamma',
-                             ]
-                        }
         if include:
             # what columns should we include/exclude?
             essential_columns = ['labels', 'species', 'multiplicity', 'tags', 'file']
@@ -666,52 +374,6 @@ def nmr(files,
     # write to file(s)
     print_results(dfs, output, output_format, verbose)
         
-def print_results(dfs, output, output_format, verbose):
-    nframes = len(dfs)
-    # rename columns to include units for those that have units
-    for df in dfs:
-        df.rename(columns=units_rename, inplace=True)
-    if output:
-        for i, df in enumerate(dfs):
-
-            if nframes > 1:
-                # then we want to write out 
-                # each dataframe to a separate file
-                # so let's prefix the filename
-                magrespath = df['file'].iloc[0]
-                prefix='_'.join(os.path.splitext(magrespath)[0].split('/')) + '-'
-            else:
-                prefix = ''
-            fname = prefix + output
-            if verbose:
-                click.echo(f'Writing output to {fname}')
-            if not output_format:
-                # try to guess format from extension
-                output_format = os.path.splitext(fname)[1][1:]
-
-            if output_format == 'csv':
-                df.to_csv(fname, index=True)
-            elif output_format == 'json':
-                df.to_json(fname)
-            else:
-                raise ValueError(f'Unknown output format: {output_format}')
-    else:
-        # We write to stdout
-
-        # if there's only one dataframe
-        # but it contains output from mutliple magres files
-        # then we need that file column
-        if nframes ==1 and dfs[0]['file'].nunique() > 1:
-            # then there's only one dataframe
-            # and we want the filename for each row
-            click.echo(dfs[0])
-        else:
-            # we can drop the file column
-            for df in dfs:
-                fname = df['file'].iloc[0]
-                click.echo(f"\n\nExtracted data from: {fname}")
-                df.drop('file', axis=1, inplace=True)
-                click.echo(df)
 
 
 def get_ms_summary(atoms, euler_convention, references, gradients):
@@ -802,136 +464,12 @@ def get_efg_summary(atoms, isotopes, euler_convention):
 
     return efg_summary
 
-def sortdf(df, sortby, sort_order):
-    ''' sort df by column, return new df'''
-
-    if sortby:
-        if sortby in df.columns: 
-            ascending = sort_order == 'ascending'
-            if sortby == 'labels':
-                isalpha = df[sortby].str.isalpha()
-                if all(isalpha):
-                    # no cif style labels, so sort by alphabetical order
-                    df = df.sort_values(by=sortby, ascending=ascending)
-                elif any(isalpha):
-                    # tricky -- we have a mix of cif style labels and non-cif style labels
-                    # so we need to sort by the cif style labels first, then sort by the non-cif style labels
-                    # this is a bit of a hack, but it works
-                    raise NotImplementedError('Sorting by labels is not implemented'
-                            'for a mix of cif and non-cif -style labels')
-                else:
-                    # all cif stlye
-                    # some care is needed to sort by labels in a natural way
-                    df[['_str', '_int']] = df[sortby].str.extract(r'([a-zA-Z]*)(\d*)')
-                    df['_int'] = df['_int'].astype(int)
-
-                    df = df.sort_values(by=['_str', '_int'], ascending=ascending).drop(['_int', '_str'], axis=1)    
-            else:
-                df = df.sort_values(by=sortby, ascending=ascending)
-        else:
-            raise ValueError(f'{sortby} not found in summary columns names')
-    return df
 
 
 
 
-def find_XHn_groups(atoms, pattern_string, tags=None, vdw_scale=1.0):
-    """Find groups of atoms based on a functional group pattern. 
-    The pattern is a string such as CH3 or CH2. 
-    It must contain an element symbol, H and the number of H atoms
 
 
-    | Args:
-    |   atoms (ase.Atoms): Atoms object on which to perform selection
-    |   pattern_string (str): functional group pattern e.g. 'CH3'
-    |                        for a methyl group. Assumes the group is
-    |                        the thing(s) connected to the first atom.
-    |                        They can be combined, comma separated.
-    |                        TODO: add SMILES/SMARTS support?
-    |   vdw_scale (float): scale factor for vdw radius (used for bond searching)
-    """
-    from soprano.properties.linkage import Bonds
-    
-    if tags is None:
-        tags = np.arange(len(atoms))
-
-    bcalc = Bonds(vdw_scale=vdw_scale, return_matrix=True)
-    bonds, bmat = bcalc(atoms)
-    all_groups = []
-    for group_pattern in pattern_string.split(','):
-        # split into central element and number of H atoms
-        if 'H' not in group_pattern:
-            raise ValueError(f'{group_pattern} is not a valid group pattern '
-                             '(must contain an element symbol, H, and the number of H atoms. e.g. CH3)')
-        X, n = group_pattern.split('H')
-        n = int(n)
-        # Find XHn groups
-        symbs = np.array(atoms.get_chemical_symbols())
-        hinds = np.where(symbs == "H")[0]
-        groups = []
-        xinds = np.where(symbs == X)[0]
-        xinds = xinds[np.where(np.sum(bmat[xinds][:, hinds], axis=1) == n)[0]]
-        # group_tags = np.ones((len(xinds), n), dtype=int)
-        seen_tags = []
-        for ix, xind in enumerate(xinds):
-            group = list(np.where(bmat[xind][hinds] == 1)[0])
-            assert len(group) == n
-            match = []
-            if len(seen_tags)>0:
-                match = np.where((seen_tags == tags[group]).all(axis=1))[0]
-            
-            if len(match) == 1:
-                # how to handle this?
-                groups[match[0]] += group
-            elif len(match) == 0:
-                seen_tags.append(tags[group])
-                groups.append(group)
-            else:
-                raise ValueError(f'Found multiple matches for {group_pattern}')
-            
-        all_groups.append(groups)
-
-    return all_groups
 
 
-def units_rename(colname, units_dict=UNITS):
-    for key, unit in units_dict.items():
-        if key in colname:
-            return f'{colname}/{unit}'
-    # if no matches found, return original name
-    return colname
 
-def get_matching_cols(df, lst):
-    """
-    Get the columns of a dataframe that roughly match a list of strings.
-    """
-    return [col for col in df.columns if any(x in col for x in lst)]
-def get_missing_cols(df, lst):
-    """
-    Get the items in list that don't match any of the columns of a dataframe
-    """
-    return [x for x in lst if all(x not in col for col in df.columns)]
-def get_duplicates(seq):
-    '''
-    Returns dict {duplicate_value: [indices]} for duplicates in a list
-    '''
-    tally = defaultdict(list)
-    for i,item in enumerate(seq):
-        tally[item].append(i)
-    return dict([(key,locs) for key,locs in tally.items() 
-                            if len(locs)>1])
-def average_quaternions_by_tags(quaternions, tags):
-    '''
-    For repeated tags, average the quaternions.
-    Return the modified list of quaternions.
-    '''
-    if len(set(tags)) != len(tags):
-        dupl_dict = get_duplicates(tags)
-        for tag, idx in dupl_dict.items():
-            quat_group = [quaternions[ig] for ig in idx]
-            # compute group average quaternion
-            quat_av = average_quaternions(quat_group)
-            # update list of quaternions
-            for ig in idx:
-                quaternions[ig] = quat_av
-    return quaternions
